Create Length_of_String.c
Question:

Implement a C program that includes a function that returns the length of a string. For example, given the string “Hello”, your function needs to return 5. Therefore, your function needs to accept a String and return an integer value. You cannot use string.h library for this exercise.
A sample run would be as follows:
Please enter a string: cngcourse
Your string has 9 characters.